+++ fastrace/src/local/local_span.rs
@@ -288,7 +288,7 @@ mod tests {
 
     fn token(trace: u128, collect_id: usize) -> CollectToken {
         CollectToken {
-            trace_id: TraceId::from_bytes(trace.to_be_bytes()).unwrap(),
+            trace_id: TraceId::from_bytes(trace.to_be_bytes()),
             parent_id: None,
             trace_flags: TraceFlags::SAMPLED,
             trace_state: TraceState::EMPTY,